requirements.txt
Django==4.2.16
django-active-link==0.1.8
django-allauth==65.1.0
python-dotenv==1.0.0
sorl-thumbnail==12.11.0
pillow==10.4.0
PyPDF2==3.0.1
openpyxl>=3.1.0
pandas==2.2.3